support non s3 config

+++ src/main/java/fr/istic/service/TemplateService.java
@@ -14,6 +14,7 @@ import fr.istic.service.mapper.TemplateContentMapper;
 import fr.istic.service.mapper.TemplateMapper;
 
 import org.apache.commons.io.IOUtils;
+import org.eclipse.microprofile.config.inject.ConfigProperty;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -41,24 +42,34 @@ public class TemplateService {
     @Inject
     TemplateContentMapper templateContentMapper;
 
+    @ConfigProperty(name = "correctexam.uses3", defaultValue = "false")
+    boolean uses3;
+
     @Transactional
     public TemplateDTOContent persistOrUpdate(TemplateDTOContent templateDTO) {
         log.debug("Request to save Template : {}", templateDTO);
         var template = templateContentMapper.toEntity(templateDTO);
-        byte[] bytes = templateDTO.content;
-        template.content = null;
-        template = Template.persistOrUpdate(template);
 
-        try {
-            fichierS3Service.putObject("template/" + template.id + ".pdf", bytes, templateDTO.contentContentType);
-        } catch (InvalidKeyException | NoSuchAlgorithmException | IllegalArgumentException | IOException e) {
-            e.printStackTrace();
+        if (this.uses3) {
+            byte[] bytes = templateDTO.content;
+            template.content = null;
+            template = Template.persistOrUpdate(template);
+
+            try {
+                fichierS3Service.putObject("template/" + template.id + ".pdf", bytes, templateDTO.contentContentType);
+            } catch (InvalidKeyException | NoSuchAlgorithmException | IllegalArgumentException | IOException e) {
+                e.printStackTrace();
+            }
+
+            TemplateDTOContent dto = templateContentMapper.toDto(template);
+            dto.content = bytes;
+
+            return dto;
+        } else {
+            template = Template.persistOrUpdate(template);
+            TemplateDTOContent dto = templateContentMapper.toDto(template);
+            return dto;
         }
-
-        TemplateDTOContent dto = templateContentMapper.toDto(template);
-        dto.content = bytes;
-
-        return dto;
     }
 
     /**
@@ -69,12 +80,10 @@ public class TemplateService {
     @Transactional
     public void delete(Long id) {
         log.debug("Request to delete Template : {}", id);
-/*        Template.findByIdOptional(id).ifPresent(template -> {
-            template.delete();
-        });*/
-        Optional<Template> templateop =Template.findByIdOptional(id);
+        Optional<Template> templateop = Template.findByIdOptional(id);
         templateop.ifPresent(template -> {
-            if (this.fichierS3Service.isObjectExist("template/" + template.id + ".pdf")) {
+            if (this.uses3) {
+                if (this.fichierS3Service.isObjectExist("template/" + template.id + ".pdf")) {
                     try {
                         this.fichierS3Service.deleteObject("template/" + template.id + ".pdf");
                     } catch (InvalidKeyException | ErrorResponseException | InsufficientDataException
@@ -84,6 +93,7 @@ public class TemplateService {
                         e.printStackTrace();
                     }
 
+                }
             }
 
             template.delete();
@@ -99,27 +109,32 @@ public class TemplateService {
      */
     public Optional<TemplateDTOContent> findOne(Long id) {
         log.debug("Request to get Template : {}", id);
-
         Optional<Template> templateop = Template.findByIdOptional(id);
-        if (templateop.isPresent()) {
-            Template template = templateop.get();
-            if (this.fichierS3Service.isObjectExist("template/" + template.id + ".pdf")) {
-                byte[] bytes;
-                try {
-                    bytes = IOUtils.toByteArray(this.fichierS3Service.getObject("template/" + template.id + ".pdf"));
-                    template.content = bytes;
+        if (this.uses3) {
+            if (templateop.isPresent()) {
+                Template template = templateop.get();
+                if (this.fichierS3Service.isObjectExist("template/" + template.id + ".pdf")) {
+                    byte[] bytes;
+                    try {
+                        bytes = IOUtils
+                                .toByteArray(this.fichierS3Service.getObject("template/" + template.id + ".pdf"));
+                        template.content = bytes;
 
-                } catch (InvalidKeyException | NoSuchAlgorithmException | IllegalArgumentException | IOException e) {
-                    e.printStackTrace();
-                }
-            } else {
-                byte[] bytes = template.content;
-                try {
-                    fichierS3Service.putObject("template/" + template.id + ".pdf", bytes, template.contentContentType);
-                } catch (InvalidKeyException | NoSuchAlgorithmException | IllegalArgumentException | IOException e) {
-                    e.printStackTrace();
-                }
+                    } catch (InvalidKeyException | NoSuchAlgorithmException | IllegalArgumentException
+                            | IOException e) {
+                        e.printStackTrace();
+                    }
+                } else {
+                    byte[] bytes = template.content;
+                    try {
+                        fichierS3Service.putObject("template/" + template.id + ".pdf", bytes,
+                                template.contentContentType);
+                    } catch (InvalidKeyException | NoSuchAlgorithmException | IllegalArgumentException
+                            | IOException e) {
+                        e.printStackTrace();
+                    }
 
+                }
             }
         }
         return templateop
